Restructure builders for encodable & decodable
